Add second benchmark file
